requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Migrate==4.0.4
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gotrue==1.0.2
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
invoke==1.7.3
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
opencv-python==4.7.0.72
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.2.0
postgrest==0.10.6
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
Pygments==2.15.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-gitlab==3.14.0
python-semantic-release==7.33.2
pytz==2023.3
readme-renderer==37.3
realtime==1.0.0
redis==3.5.3
redisearch==2.1.1
rejson==0.5.6
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rq==1.13.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.10
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
yarl==1.9.2
zeep==4.2.1
zipp==3.15.0
